Remove `BuiltinLintDiag::NamedArgumentUsedPositionally` variant
